Update Status of MIP12c2-SP2
Also merging this proposal and updating the status to `Obsolete` as the proposal falls under the last point in the `Obsolete` criteria in MIP0, "A MIP no longer makes sense to keep in consideration.". This is because this proposal was rejected from the governance cycle due to missing work requirements and instead went through the weekly cycle (final vote for KNC is June 26th). The weekly cycle is being used until we can clean up and clarify the format for proposing MIP12 subproposals and figure out what exactly needs to be included in MIP12 subproposals going forward.