Fix accidental omission in the Atlas
Ratified amendment subproposal MIP102c2-SP34 amended several Scopes and the Atlas. It is a general edit amendment, i.e., it cannot modify Article 1s in Scopes; however, there is no such restriction for the Atlas itself. When we merged the associated PR, we accidentally omitted the Article 1 amendment for the Atlas.

We're hereby fixing  this mistake.